requirements.txt
# Python >= 3.10 required for tsapython
# To install tsapython from GitHub (if PyPI fails):
# pip install git+https://github.com/LC-Linkous/tinySA_python.git#subdirectory=tsapython

tsapython>=2.0.0
pyserial>=3.5
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0
matplotlib>=3.5.0

# Optional: for FITS file support
# astropy>=5.0.0
